requirements_setup.txt
pydantic            >= 2.0,     < 3.0
pydantic-settings   >= 2.0,     < 3.0
ruamel.yaml         >= 0.17,    < 0.19
typing-extensions   >= 4.4,     < 5
